나인볼 매치 카운트 업데이트중 형변환 오류 수정

+++ app/src/main/java/wook/pool/board/data/repository/CountRepository.kt
@@ -24,7 +24,7 @@ class CountRepository @Inject constructor(
 
     fun updateNineBallMatchTotalCount(
             documentPath: String,
-            count: Int,
+            count: Long,
             onSuccess: () -> Unit,
             onFailure: (Exception) -> Unit
     ) {
